+++ fighters/common/src/function_hooks/energy/motion.rs
@@ -280,7 +280,8 @@ unsafe fn motion_update(energy: &mut FighterKineticEnergyMotion, boma: &mut Batt
         *FIGHTER_STATUS_KIND_ATTACK_LW3])
     {
         let mut stop_energy = KineticModule::get_energy(boma, *FIGHTER_KINETIC_ENERGY_ID_STOP) as *mut app::KineticEnergy;
-        let reset_speed_2f = smash::phx::Vector2f { x: energy.prev_speed.x, y: energy.prev_speed.y };
+        let prev_speed = KineticModule::get_sum_speed3f(boma, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_MAIN);
+        let reset_speed_2f = smash::phx::Vector2f { x: prev_speed.x, y: prev_speed.y };
         let reset_speed_3f = smash::phx::Vector3f { x: 0.0, y: 0.0, z: 0.0 };
         lua_bind::KineticEnergy::reset_energy(stop_energy, *ENERGY_STOP_RESET_TYPE_GROUND, &reset_speed_2f, &reset_speed_3f, boma);
         lua_bind::KineticEnergy::enable(stop_energy);
