+++ fighters/common/src/general_statuses/attack/attack.rs
@@ -2,7 +2,7 @@
 use super::*;
 use globals::*;
 
-// This file contains code for wavelanding
+// This file contains code for jabs
 
 pub fn install() {
     skyline::nro::add_hook(nro_hook);
