+++ fighters/common/src/general_statuses/attack/mod.rs
@@ -1,0 +1,10 @@
+use super::*;
+mod attack;
+mod attackx4;
+mod attackdash;
+
+pub fn install() {
+    attack::install();
+    attackx4::install();
+    attackdash::install();
+}
