+++ fighters/common/src/general_statuses/mod.rs
@@ -13,13 +13,10 @@ mod jumpsquat;
 pub mod jump;
 mod footstool;
 mod run;
-pub mod attack;
+mod attack;
 mod shield;
 mod turn;
 mod walk;
-mod attackdash;
-mod attackhi4;
-mod attacklw4;
 mod passive;
 mod damagefall;
 mod downdamage;
@@ -536,9 +533,6 @@ pub fn install() {
     shield::install();
     turn::install();
     walk::install();
-    attackdash::install();
-    attackhi4::install();
-    attacklw4::install();
     passive::install();
     damagefall::install();
     downdamage::install();
